Removed login file and moved the code to App.js where it makes more sense to be.
