+++ package.json
@@ -9,7 +9,8 @@
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "3.4.1"
+    "react-scripts": "3.4.1",
+    "spotify-web-api-js": "^1.4.0"
   },
   "scripts": {
     "start": "react-scripts start",
